nit: split test/utils.go => test/{addrs,errors}.go.
