+++ test/addrs.go
@@ -7,20 +7,6 @@ import (
 	ma "github.com/multiformats/go-multiaddr"
 )
 
-func AssertNilError(t *testing.T, err error) {
-	t.Helper()
-	if err != nil {
-		t.Errorf("unexpected error: %v", err)
-	}
-}
-
-func ExpectError(t *testing.T, err error, msg string) {
-	t.Helper()
-	if err == nil {
-		t.Error(msg)
-	}
-}
-
 func GenerateTestAddrs(n int) []ma.Multiaddr {
 	out := make([]ma.Multiaddr, n)
 	for i := 0; i < n; i++ {
